+++ labour/views/admin_views.py
@@ -92,68 +92,3 @@ def labour_admin_mail_view(request, vars, event):
     )
 
     return render(request, 'labour_admin_mail_view.jade', vars)
-
-
-@labour_admin_required
-@require_http_methods(['GET', 'HEAD', 'POST'])
-def labour_admin_mail_editor_view(request, vars, event, message_id=None):
-    from mailings.models import Message
-    from mailings.forms import MessageForm
-
-    if message_id:
-        message = get_object_or_404(Message, recipient__event=event, pk=int(message_id))
-    else:
-        message = None
-
-    form = initialize_form(MessageForm, request, event=event, instance=message)
-
-    if request.method == 'POST':
-        if 'delete' in request.POST:
-            assert not message.sent_at
-            message.delete()
-            messages.success(request, 'Viesti poistettiin.')
-            return redirect('labour_admin_mail_view', event.slug)
-
-        else:
-            if form.is_valid():
-                message = form.save(commit=False)
-
-                if 'save-send' in request.POST:
-                    message.save()
-                    message.send()
-                    messages.success(request, 'Viesti lähetettiin. Se lähetetään automaattisesti myös kaikille uusille vastaanottajille.')
-
-                elif 'save-expire' in request.POST:
-                    message.save()
-                    message.expire()
-                    messages.success(request, 'Viesti merkittiin vanhentuneeksi. Sitä ei lähetetä enää uusille vastaanottajille.')
-
-                elif 'save-unexpire' in request.POST:
-                    message.save()
-                    message.unexpire()
-                    messages.success(request, 'Viesti otettiin uudelleen käyttöön. Se lähetetään automaattisesti myös kaikille uusille vastaanottajille.')
-
-                elif 'save-return' in request.POST:
-                    message.save()
-                    messages.success(request, 'Muutokset viestiin tallennettiin.')
-                    return redirect('labour_admin_mail_view', event.slug)
-
-                elif 'save-edit' in request.POST:
-                    message.save()
-                    messages.success(request, 'Muutokset viestiin tallennettiin.')
-
-                else:
-                    messages.error(request, 'Tuntematon toiminto.')
-
-                return redirect('labour_admin_mail_editor_view', event.slug, message.pk)
-
-            else:
-                messages.error(request, 'Ole hyvä ja tarkasta lomake.')
-
-    vars.update(
-        message=message,
-        form=form,
-        sender="TODO",
-    )
-
-    return render(request, 'labour_admin_mail_editor_view.jade', vars)
